Change user.id to int64

+++ professional-service/internal/domain/user/user_model.go
@@ -1,8 +1,7 @@
 package user
 
 type User struct {
-	// TODO: convert it to int64
-	ID           int     `json:"-"`
+	ID           int64   `json:"-"`
 	FirstName    string  `json:"firstName,omitempty"`
 	LastName     string  `json:"lastName,omitempty"`
 	Email        string  `json:"email,omitempty"`
